Adds a Cas2v2UserEntity
Adds a new entity for Cas2v2 users where all user types are represented
in a single entity (Cas2v2UserEntity) and differentiated by the
userType field which is a Cas2v2UserType containing the different
allowed user types for Cas2v2.

As a result of the change, all of the Cas2v2 classes that contain users
have been changed to use the new type instead of NomisUserEntity/UserEntity/ExternalUserEntity.
This removes a dependence on the Cas2 entities but should still allow a
future migration to use the entities in this namespace.

This PR also introduces a CAS2V2 ServiceName which is used to ensure
that a Cas2v2Application is used instead of Application or Cas2Application
when calling APIs.

DomainEvents remain as cas2 domain events with no change