Tidy up ‘seed on startup’ config
Before this commit the terminology ‘auto’, ‘auto-seed’ and ‘auto-script’ were used to configure and describe seed jobs and scripts that run on on startup. This commit updates the configuration to instead use the terms ‘seed on startup’, removing the use of ‘auto’. It also re-organises the configuration to give it a more logical hierarchical structure. The configuration itself for each environment is unchanged.