Split ‘dev+test’ migrations
This commit splits the ‘dev+test’ migrations into separate folders, ensuring we do not truncate any table in test on deploy

This commit also reintroduces envronment-specific config in the test helm config. This was previously removed to ensure elements are not cleared down in test on deploy, as a temporary workaround.