Update Bed Search Filters
This updates the BedSearchParameters model to TemoraryAccommodationBedSearchParameters (renaming to come later as this will affect the UI), as only cas3 uses it now. It also refactors the controller and service to pass the filter object through instead of individual parameters, this is initial refactoring in order to maintain this endpoint during refactoring to add new filter options.