Add /cas1/reference-data/departure-reasons
This should be used instead of the service-agnostic endpoint, which can return unexpected results due to the amiguous SQL being used (this is documented in this commit).

Instead of ‘fixing’ the amiguous SQL and potentially disrupting CAS3, we’ve instead introduced a CAS1 specific version of the SQL/endpoint

This commit also ‘splits’ a wildcarded departure reason to instead have CAS1 and CAS3 specific reasons. The existing reason was change to belong to CAS3 to avoid disrupting CAS3. This allows us to disable the reason for CAS1 without affecting CAS3