spike/CAS-1341-CAS2-Prison-Transfer-Event-Emit-POC
This is a POC branch to demonstrate how to set up a listener for the domain events topic.

Message Listener listens to the inboundQueue which is filtered by the allocation and handover event types and prints out the message.

You can run it locally and send messages using the scripts found in test-messages in the test resources folder.

It has a simple Integration test that checks that the listener was called and also a tests to check the health of the queue.