updated role check on applications
Application endpoint should only be called by CAS3 Refferers, so assessors should never be hitting this endpoint. Added validation so if people don't have the referrer role, they should not be able to access it.

Chunked the calls to delius into batches of 500, as this will throw an error if the person has 501+ open applications (unlikely outside of dev/test)

refactored call to getExistingUserOrCreateDeprecated to use getExistingUserOrCreate