Update hmpps.gradle-spring-boot
Upgrade to the latest hmpps.gradle-spring-boot plugin, ensuring we ignore irrelevant vulnerability CVE-2024-50379 during trivy scans.

This will upgrade spring boot from 3.4.0 to 3.4.1. Hibernate Spatial has been upgraded to align with the newer version of hibernate used in 3.4.1 (note that it previously not aligned with the spring boot hibernate version)