Enable tier-api retry on timeout
This commit enables retry on timeout for tier-api and reduces its timeout threshold from 10 seconds to 2 seconds. Observations have shown that when we do get a timeout and a 500 is returned to the UI a retry triggered by the UI succeeds.

The timeout of 2 seconds is based upon the 99th percentil response time of 47 milliseconds, with only one request taking less than 1 second (without timeout) in January so far.