Run CAS2 Bail e2e tests after deploying to dev
As we have a new CAS2 Bail repo making use of the API (namespace
`CAS2V2`) we want to add the Bail e2e tests to the list of tests that
are run before deploying to pre-prod to avoid regressions.

As no pre-prod or prod environments are currently in use for CAS2 Bail,
this step is not required to pass before deploying to pre-prod, but will
be the case in the future.