Add arson offence characteristic
This characteristic will be associated with any placement request that was created from an assessment where ‘arson offences’ was selected. This is currently represented by the isArsonSuitable characteristic, but that mapping is incorrect and will be repurposed in subsequent changes.

This characteristic is disabled by default because it will not be added to any placement requests once the aforementioned changes have been made