CAS1 App Report Assesor Bug Fix
Before this commit we retrieved the ‘initial assessor’ information from the assessment decision domain event. This meant that no information was shown about the assessor until a decision had been made.

This commit changes the report to instead retrieve assessor information from the user’s table, using the value of `assessment.allocated_to_user_id`. Ideally we’d use the assessment allocation domain event, but we do not currently capture the assessor’s CRU on the domain event.

We’ve checked to ensure that the allocated user is not removed from the assessor on submission

```
select count(*) from assessments where decision IS NOT NULL and allocated_to_user_id IS NULL and service != 'temporary-accommodation';
-- success --
count
-----
    0
```

The only assessments without an allocation are those pending allocation

```
select count(*) from assessments where allocated_to_user_id IS NULL and decision IS NULL and reallocated_at is null and service != 'temporary-accommodation';
-- success --
count
-----
  118
```