Retain non-redundant decimal points in sitesurveys
When 'general' types are read by Dataframe they're assumed to be Doubles. If converted to a String, a decimal place will always be present e.g. '1' in Excel will appear as '1.0'. To avoid this issue the new code strips out redundant decimal places whilst retaining non 0 decimal places.

This is an imperfect solution because if a site survey legitimately included a numeric value of 1.0, this would be truncated to '1'. In this case we could look at using POI directly which doesn't have that issue (see example code on APS-1933).

We have checked all site surveys and non are using `1.0` or other similar redundant decimal place