Local seeding fails if many existing applications
The local seed job fails on startup with an out of memory exception if there are many existing applications in the database. This is becuase the logic was retreiving all applications for a given CRN to check if any exists. The logic is now changed to do that check using SQL