Reduce SQL used to build CAS1 space planning model
Before this commit the following code in the `SpacePlanningModelsFactory` was used to collect all beds in a given premise;

```
premises.rooms.flatMap { it.beds }
```

This trigger a lazy load for each room, and a lazy load for all beds in each room

This commit updates the `SpacePlanningModelsFactory` to instead use a single native SQL query to retrieve all beds in a single query.

This commit also simplifies characteristic filtering when building models because it’s safe to assume that any characteristic in the  `ROOM_CHARACTERISTICS_OF_INTEREST` list is active and scoped to rooms.