Remove OSIV from /cas1 endpoints
Open Entity (Session) In View (OSIV) is enabled by default in Spring, and whilst simplifying development wrt. lazy loading JPA relationships, it results in a connection being held open from it’s first usage to completion of the API request. This can lead to empty connection pools under load, especially when calls are being made to upstream services in the same thread that may take some time to respond.

Because removing OSIV can cause subtle issues (especially are object equality), we’re only applying it to the /cas1 endpoints in this commit, most of which are not yet in use in production.