Lazy load entity schema versions
This commit makes schemaVerson relationships lazy loaded. These are only required when updating/submitting an application/placement application/entity, yet they are eagerly loaded on every retrieval of the entity, something which happens far more often.

Note that because lazy-loaded ‘proxies’ now exist in the open session for `JsonSchemaEntity`s, these are sometimes returned when getting a specific schema via JPA. For that reason we have to wrap such gets in a call to `Hibernate.unproxy` to ‘unwrap’ it