Add filtering to bedspace search
This PR updates how filtering works on bed searches. You can now pass in a list of UUIDs for premises and bedspace (room), which can either be included or excluded. This is because we want to see beds that ARE wheelchair accessible, but ARE NOT suitable for sexual offenders.

There is going to be ongoing refactoring work which will remove the mapping from characteristicId to propertyName, and that will be removed in a future PR.