Disable CAS1 overbooking summary
The overbooking summary shown on the view premise page causes performance issues when under load. This is for a few reasons

* calculating the summary requires pulling data from many tables and some ‘in memory’ calculations across this data
* the summary has to consider 3 months of bookings data

We have decided to disable this summary for now (by always returning ‘0’ overbookings), resolving performance issues. If this is required we should consider caching this result and invalidating the cache when changes are made that may affect the overbooking calcluation (e.g. premise/room definitions, booking created/amended/cancelled, overbooking created/amended/cancelled).