Add Cas1SpaceBooking.placementRequestId
When the `Cas1SpaceBooking` API model was originally created a field `requestForPlacementId` was provided that aligned with the `requestForPlacement` endpoints.

The UI does not use these endpoints, and instead always wants to retrieve the associated request for placement, if defined. Given this, the `requestForPlacement` field has been deprecated and updated to only provide requestForPlacementId to fix the immediate problem. A new `placementRequestId` field has been added to clarify the value being provided.