Use space bookings to test placement req outcome
Before this commit we were using domain events created for legacy bookings when testing the Placement Matching Outcomes report. This commit changes the test code to instead use space bookings. Whilst this doesn’t create domain events that are functionally any different, it removes usage of code that will be, in time, deprecated.