Add withdrawn entries to placement matching report
Before this commit we excluded withdrawn requests for placement from the placement matching outcome report. Because any ‘not matched’ request for placment is withdrawn once the arrival date has passed, this meant we never saw ‘not matched’ entries in historical reports.

This commit now included entries where the request for placement was subsequently withdrawn