Deprecate `Cas1SpaceSearchRequirements`
To simplify the space search API we want to remove `Cas1SpaceSearchRequirements`. Before this commit only one of the 4 properties in this class were being used.

This commit deprecates all fields on `Cas1SpaceSearchRequirements` and introduces a new `spaceCharacteristics` property on `Cas1SpaceSearchParameters` to replace the deprecated field

The deprecate fields continue to operate as before.