Rename Mental Health Premise Characteristics
This commit renames the premises-specific characteristics related to mental health to align with existing entries in the `PlacementRequirements` API Type.

This ensures that going forward we’ll capture these characteristics in the `placement_requirements` table when an assessment is submitted because corresponding characteristics now exist in the database (previously these were ignored because a related characteristic couldn’t be found).