Partial match arson suitable site survey question
Was

`Is this room suitable for people who pose an arson risk? (Must answer yes to Q; 6 & 7, and 9 or  10)`

Now

`Is this room suitable for people who pose an arson risk? (Must answer yes to Q; 4, 6 & 7, and 9 or  10)`

Using a partial match on the question will make it easier to deal with minor changes to questions going forward