Remove use of deprecated func in CAS3 migrate job
Note that the integration was returning in an incorrect response for calls to /probation-cases/summary. It should have returned an empty list, but was instead returning a 404 (because no CRNs were configured in wiremock for the get params). The test still worked because the now deprecated `OffenderDetailsDataSource` misinterpreted a 404 as ‘offender not found’ which was incorrect.