OOSB report input optimisation
Only consider beds with at least 1 out of service bed records when building the out of service beds report.

This considerably decreases the size of the input to the report. The report itself is still inefficient and would be better implemented via SQL that only considers records within the given time frame, instead of starting with a large set of beds and then checking each one to see if there’s an OOSB within the given time frame