Use internal enum for PlacementRequest Gender
Before this commit we were using an enum generated from API specs to map Gender on Placement Request. This is a problem because we’re also using ordinal based mapping in the database, instead of name based mappings. This means if the ordering of the generated enum changes, and subsequent retrievals from the database could be wrong.

This commit doesn’t fix the ordinal mapping issue, but does introduce an internal ‘manually’ managed enum with a warning comment making it less likely for the ordering of the enum to be changed.

For reference the generated API type is as follows:

```
enum class Gender(@get:JsonValue val value: kotlin.String) {

    male("male"),
    female("female");

    companion object {
        @JvmStatic
        @JsonCreator
        fun forValue(value: kotlin.String): Gender {
                return values().first{it -> it.value == value}
        }
    }
}
```