Add DomainEventType.emittable config
This commit adds a new boolean property `DomainEventType.emittable` used to indicate if any domain events of this type are emittable. This is currently only used by CAS1.

This commit updates the logic in `Cas1DomainEventService` to observe this logic and updates `APPROVED_PREMISES_REQUEST_FOR_PLACEMENT_CREATED.emittable` to false, to give us a domain event type to use in our tests.

Subsequent commits will set `emittable` to true for all domain event types that are never currently emitted