Add placement_request_id to domain_events
With the introduction of cas1 change requests it’s likely we’ll need a way to group domain events relates to a given placement requuest. We currently use Domain Event Meta data to do this reporting, but it’s not an efficient way to query the data.

This commit adds a new indexed column `domain_events.cas1_placement_requests` to resolve these issues