Run prod to preprod restore earlier in the day
The job currently runs at 9am british summer time. It’s more helpful to have this job run before devs start work (e.g. before 9am) as it temporarily makes pre-prod unavailable. It also means if the job fails to run an alert will be available for us to respond to when we start work.