Limit timeline applications loaded into memory
Whilst we limit the number of applications returned to the timeline to 50, before this commit we were still loading them into memory.

This commit ensures we only ever load a maximum of 50 applications (and 50 offline applications) into memory when loading the timeline.

The solution isn’t perfect - we could theoretically load 50 regular applications and 50 offline applciations into memory, despite only returning 50 of those combined. Given this logic is a fix to handle large timelines when running performance tests reusing the same CRN, this is deemed acceptable.