Check in CAS1 domain event generated code
Add the CAS1 domain event generated code and remove build configuration to generate it from yml. Also remove the yml.

This commit also updates the ktlint and detekt configuration to not check the generated code as it’s not trivial to ‘bring it up to spec’

TODO: will it still format it?

f