Remove CAS1 Domain Events API YML
This commit removes the CAS1 Domain Events API YML, checking in the generated code.

This allows us to convert the `Cas1DomainEventsController` intro a fully fledged controller, instead of a delegate.

This was achieved by temporarily disabling the `put("delegatePattern", "true”)` in build.gradle to give us a simple controller implementation to copy/paste.