Remove CAS1 Domain Events API YML
This commit removes the CAS1 Domain Events API YML, checking in the generated code.

This change has allowed us to convert the `Cas1DomainEventsController` intro a fully fledged controller, instead of a delegate. Simplifying the code, making it more readible and allowing us to better leverage reuse/interfaces in the API models.

To simplify updating the controller, we temporarily disabled the delegate pattern in the `build.gradle` file e.g. `put("delegatePattern", "true”)`. This gave us a simpler controller implementation to merge into our existing controller code. I also had to manually format the controller to ensure the line length remained within the detekt limits.

We are ignoring any detekt/ktlint issues with the API models as it would be too time confusing to ‘fix’ the models to match coding standards (Specifically, updating all the enum definitions to be uppercase).