Add `cas1_change_requests.resolved`
This commit adds a boolean `resolved` to the change request entity, used to indicate if a change request is resolved. This is requried to support ‘closing’ a change request if the corresponding placement is cancelled, or in a state such that the change request is no longer relevant.

A change request will be resolved if

- It’s accepted
- It’s rejected
- The corresponding placement is in a state such that the change request is no longer relevant

This commit also removes the `decision made at` value and replaces it with `resolved at`, which will be the same date/time.