Introduce common domain event envelope interface
This commit updates all CAS1 domain event envelopes to use a common interface. This allows us to then define the envelope type associated with each domain event type on `DomainEventType`, ensure the referenced type is actually an envelope. We can then use this new configuration in the `Cas1DomainEventService` when unmarshalling domain event json without the need for a large, difficult to read `when` structure.

This helps us move towards to moving all high-level domain event configuration into the `DomainEventType`, making it much easier to understand what is required when adding a new domain event