Use generic domain event envelope for CAS1 GET
We currently define an envelope type for each individual domain event, despite the envelope always containing the same 4 fields (as defined by the `Cas1DomainEventEnvelopeInterface`). This was previously required when defining domain event schema in YML to differentiate between the different types.

This commit introduces a generic envelope (`Cas1DomainEventEnvelope`) that can be used for all domain events, removing the need to add individual envelope types. It updates all GET requests to use this generic envelope, and deprecates all bespoke envelopes.

Subsequent commits will use the envelope when persisting domain events, which should unlock removing some additional duplication of logic when building the envelope. Because this will be quite a bit of effort, this change will be made over time.