Use generic domain events for change requests
Before this commit we had separate domain events for the creation, acceptance and rejection of change requests. This was going to result in many similar domain events being created, which required additional code in the UI to render.

We have agreed that sufficient information on the outcome of the domain events will be captured on the corresponding domain event types (e.g. booking made, booking changed), as such we can use generic domain events for recording change request changes in the timeline.