Move schemaVersion on domain event timeline API
Before this commit schemaVersion was part of the domain-event-specific content payload. Because all domain events have a schema version, this has been moved to the `Cas1TimelineEvent`, reducing the amount of copy-paste code. It has been temporarily retained for the Booking Amended payload type because this field is currently in use by the UI