Add emergency transfer arrived emails
Emails are sent on arrival for any space booking that was created for an emergency transfer.

This commit introduces the user of spring events to ‘notify’ the email service that an arrival event has occurred. This minimises the amount of mocking required in the main service functions and better decouples services, simplifying unit testing