Refactor shorten logic to raise domain event
This commit refactors the shorten space booking logic to ensure a booking changed domain event is raised.

To do this it updates the `Cas1SpaceBookingService.shortenSpaceBooking` function to reuse the existing `updateSpaceBooking` function as much as possible, including raising emails and domain events. This requires a change to the `Cas1BookingEmailService` to provide a single ‘booking amended’ function for shortened bookings or otherwise.