Remove DomainEventEntity.cas1PlacementRequestId
This was added to provide an efficient way to retrieve domain events related to a given placement request. Subsequent work has identified an equally efficient way to do this using existing metadata (see notes on APS-2231).

This commit only removes the column from the entity model.  A subsequent commit will remove the column from the database. This ensures that there are no data-model-mismatch issues during rollout in prod.