Enforce validation before creating a space booking
This commit introduces a `ValidatedCreateBooking` that is now required to create a new space booking. This ensures the booking request is validated before being created.

The `ValidatedCreateBooking` contains the space booking that is going to be created. This is helpful because it provides us with access to the pre-persisted space booking, including its ID, which is required to fully populate related domain events (this is the primary motivator for this change)