Enable email in test, disable in dev
We are reconfiguring test email addresses to use a shared inbox `cas1dev@justice.gov.uk` that will be added to the notify guest list. To avoid spamming, we’re going to disable email in dev (where e2es are ran) and enable email in test where we can manually test email changes.