Fix PreemptiveCacheTest
This test has been broken for a long time, but it wasn’t been ran automatically (see previous commit).

The test was broken because we introduced automatic retries on 500 errors, which changed the number of requests being made to wiremock in this test. The test has been updated to use a 400 error instead, which is not retried.

This commit also refactors the test for readibility