Limit bookings to 2 years duration
Modify the booking create and update validation logic to ensure a booking is never more than 2 years old.

This upper limit is being set as we’ve discovered memory issues when amending bookings with a very long duration (e.g. 100 years). This is due to the capacity calculation logic trying to calculate the capacity in the premise for the length of the entire bookings. Furthermore, bookings will never be longer than 2 years, so this is a sensible limitation to put in place.