Remove bed_move entries on space booking migration
When migrating a booking to a space booking, the existence of a corresponding entry in `bed_moves` blocks deletion of the `booking` entry. Usage of bed_moves was removed in 08/24 as we stopped tracking bed allocation in CAS1. This information doesn’t need migrating to space bookings. Bed moves were only ever captured for the NE region.

This commit updates the booking to space booking migration logic removing any linked entries in bed_moves. This follows the established pattern of removing any information related to the legacy booking entry.

This commit also deprecates the bed move table, in preparation for its removal once all CAS1 bookings have been migrated to space bookings.