Simplify premises summary test
Remove the bed count test and instead check the bed counts in other tests. The previous approach could led to test failures if tests ran in a specific order because the bed count test was adding beds to premises that then ‘broke’ the bed count check in other tests.

Also simplify test names