Improve non existant CRN handling for access check
This commit updates the `canAccessOffender` function to return a null Boolean value if no record if found for a given CRN when checking if access is allowed. Previously it returns `false` (e.g. access isn’t allowed).

Any existing callers have been modified to maintain the existing behaviour.

This change is being made to support scenarios where we want to return a 404 to the caller if no access information can be found for a given CRN (i.e. the offender doesn’t exist)