Remove schema validation logic
This commit removes the schema validation logic, instead always returning that the schema is valid. This is the first step in removing the schema validation code, which has been unused for almost two years.

This intermediary steps allows us to remove the `com.networknt:json-schema-validator` dependency, which was out of date.