Bug fix for CAS-1725
This bug was causing emails to be sent to old POMs when NACRO users were leaving notes. CAS2 Bail applications may not be assigned to a prison or a POM, as court users are not assigned to a prison. This solution first checks for a POM email, if there is no POM email it will fall back to a OMU email, and if there is no OMU email, it will fall back to using the created by user email. This currently comes from the NOMIS users table, or the Cas2Users table if it does not exist there.