Increase CAS1 upcoming booking threshold
This commit updates the upcoming booking threshold from 2024-06-01 to 2025-01-01. This is in preparation for onboarding of all remaining premises to find and book, which will require an import of arrival/departure from Delius. This will include several bookingsd before 2025 that haven’t been marked as arrived or departed, despite clearly being in the past. As explained in the documentation for the constant `Cas1SpaceBookingRepository.UPCOMING_EXPECTED_DEPARTURE_THRESHOLD`, we don’t want these to appear as upcoming to avoid substantial data cleansing activities.