Truncate capacity requests to less than 2 years
We have observed heap memory issues when capacity information is requested for many years. This shouldn’t happen under normal operation at CAS1 bookings should not be lasting years, but it is possible if an expected arrival or departure date is mistyped.

To avoid memory issues in such scenarios, the capacity endpoint now transparently limits all responses to a maximum of two years