Make CAS1 OASys Metadata Endpoint future proof
This commit updates the cas1 metadata endpoint to return a container type, allowing us to add additional metadata to the response in the future (e.g. data record created/updated) without making a breaking change. The path of the endpoint has also been changes to reflect this.

This change is safe because the endpoint is not currently being used