Return CAS1 OAsys Metadata if assessment not found
By default a 404 will be returned by the CAS1 Metadata endpoint if an assessment can’t found. This commit introduces a feature-flagged behaviour that will instead return a logically empty response that indicates that an applicable assessment isn’t available.

This is part of a change to allow us to remove questions definitions from the UI, currently used when an assessment isn’t returned (indicated by a 404)