Use new questions for NOD-1057 assessments
For new assessments in OASys some questions have been removed, and some new questions added. This commit changes the `OASysSectionsTransformer` to use those new questions for new assessments. There isn’t a concrete way to determine if an assessment is ‘new’, so we use the absence of the previous questions answers to help determine this.

This behaviour only applies to CAS1 (and CAS3, but that doesn’t use the new/old questions). It is feature flagged and currently disabled in production.