CAS-1704/Add migration job for void bedspace cancellations
This PR add's new fields to the Cas3VoidBedspace entity for cancellation date, cancellation notes, and bedspace id. This enables the removal of the Cas3VoidBedspaceCancellations table, as it's a 1-1 mapping, and also adds the link to the new Cas3Bedspace models, created as part of the larger migration work.