Remove workflow manager from get offline app
Update the logic to get an offline application to use a permission instead of checking specific user roles. Add this permission to those user roles that could already access this functionaltiy, and also add CRU MEMBER, which is the logical replacement for WORKFLOW_MANAGER.

All users in prod who currently have the WORKFLOW_MANAGER role are already a CRU_MEMBER or FUTURE_MANAGER, both of which can access this endpoint.