Add job to fix incorrectly linked Placement Dates
We maintain two (redundant) links from a placement application to placement requests

 1. Placement Application <-> Placement Request
 2. Placement Application <-> Placement Date <-> Placement Request

Both (1) and (2) should refer to the same set of placement requests

Some Placement Requests in (1) are different to those referenced by (2). Investigations has shown in the case of (2) that the placement requests referred to have been reallocated, and the new versions of hte placement requests are correctly linked in (1). This job fixes the references in (2) to match those in (1)