Add `ensureUserCanAccessOffender` func
This helps reduce usage of deprecated offender service functions, and removes duplication of this logic

Three controllers that previously retrieved the offender and access details now use this function, meaning they now only retrieve access details, reducing calls to probation integration

The behaviour of Documentcontroller is now slightly different. If an access record can’t be found it’ll return a forbidden error instead of a 404. This is an improvement because a 404 implies the document can’t be found.